+++ ai-logic/firebase-ai/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-- [changed] Upgraded `isThought` and `thoughtSignature` to be public on the `Part` interface and exposed primary constructors for all `Part` subclasses to support complete object reconstruction.
+- [changed] Upgraded `isThought` and `thoughtSignature` to be public on the `Part` interface and exposed primary constructors for all `Part` subclasses to support complete object reconstruction. (#8339)
 
 # 17.13.0